From: Jeff King <peff@peff.net>
When sha1_loose_object_info() finds that a loose object file cannot
be stat(2)ed or mmap(2)ed, it returns -1 to signal an error to the
caller.  However, if it finds that the loose object file is corrupt
and the object data cannot be used from it, it forgets to do so.

This can confuse the caller, who mistakenly thinks that there is a
loose object and possibly learns an incorrect type and size from the
function.  The SHA-1 collision detection codepath, for example, when
it receives an object over the wire and tries to see if the data is
the same as what is available as a loose object locally, can get
confused to think the object that locally exists have a type or size
that is different from what came over the wire, declaring that it
encountered an SHA-1 collision, when the loose object we have
locally is corrupted.